refactor: quib -> post
