--- backend/apps/quib/api/v1/urls.py
@@ -1,8 +1,0 @@
-from rest_framework.routers import DefaultRouter
-
-from .viewsets import QuibViewSet
-
-router = DefaultRouter()
-router.register(r'', QuibViewSet)
-
-urlpatterns = router.urls
